requirements.txt
faster-whisper>=0.9.0
torch>=2.0.0
numpy>=1.22.0
pydub>=0.25.1
tqdm>=4.65.0
click>=8.1.3
joblib>=1.2.0
ffmpeg-python>=0.2.0
pytest>=7.0.0
pytest-cov>=4.0.0 